+++ GameTest/src/states/slime/SlimeMovingState.cpp
@@ -103,7 +103,7 @@ void SlimeMovingState::Update(float deltaTime) {
 	}
 
 	// Calculate difference between slime and player on X axis, only chase if <= 5 tiles
-	int diff = std::abs(player->GetX() - slime->GetX());
+	float diff = std::abs(player->GetX() - slime->GetX());
 
 	if (diff < TILE_SIZE * 5) {
 		slime->ChangeState(State::SLIME_CHASING);
